refactor PluginService
